+++ src/App.js
@@ -1,9 +1,8 @@
 import logo from './logo.svg';
 import 'bootstrap/dist/css/bootstrap.min.css';
-// import 'bootstrap/dist/js/bootstrap.bundle.min'
 import './App.css';
 import {Fragment} from "react";
-import Homepage from "./components/homepage";
+import Homepage from "./components/homepage/homepage";
 
 function App() {
   return (
